@@ -28,69 +28,69 @@
         "build_linux": "react-app-rewired build && ./CopyBuild.sh",
         "test": "react-app-rewired test",
         "eject": "react-scripts eject",
-        "build-electron": "tsc -b tsconfigElectron.json",
         "build-electron-store": "tsc -b tsconfigElectron.json && mv buildElectron/mainStore.js buildElectron/main.js",
-        "start-electron": "npm run build-electron && electron .",
-        "dist-snap": "npm run build-electron-store && rm -f dist/*.snap && electron-builder -l snap -c electronBuilderConfigs/snap.json --x64",
+        "start-electron": "npm run build-electron-store && electron .",
+        "eb-dist-snap": "npm run build-electron-store && rm -f dist/*.snap && electron-builder -l snap -c electronBuilderConfigs/snap.json",
+        "eb-dist-snap-arm64": "export SNAPCRAFT_BUILD_ENVIRONMENT=lxd && npm run build-electron-store && rm -f dist/*.snap && electron-builder -l snap -c electronBuilderConfigs/snap.json --arm64",
         "dist-snap-arm64": "export SNAPCRAFT_BUILD_ENVIRONMENT=lxd && npm run build-electron-store && rm -f dist/*.snap && electron-builder -l snap -c electronBuilderConfigs/snap.json --arm64",
+        "dist-snap": "mkdir -p dist && snapcraft snap --use-lxd -o dist",
         "publish-snap": "snapcraft upload --release=stable dist/*.snap",
         "icons-mas": "cd buildElectron && iconutil -c icns icon.iconset",
-        "dist-mas": "npm run build-electron-store && electron-builder -m mas --universal -c electronBuilderConfigs/mas.json",
-        "dist-mas-dev": "npm run build-electron-store && electron-builder -m mas-dev --universal -c electronBuilderConfigs/masDev.json"
+        "dist-mas-dev": "npm run build-electron-store && electron-builder -m mas-dev --arm64 -c electronBuilderConfigs/masDev.json",
+        "dist-mas": "npm run build-electron-store && electron-builder -m mas --universal -c electronBuilderConfigs/mas.json"
     },
     "dependencies": {
         "electron-window-state": "^5.0.3"
     },
     "devDependencies": {
-        "@ionic/react": "^5.6.12",
-        "@ionic/react-router": "^5.6.12",
-        "@testing-library/jest-dom": "^5.14.1",
-        "@testing-library/react": "^11.2.7",
-        "@testing-library/user-event": "^12.8.3",
+        "@ionic/react": "^6.0.16",
+        "@ionic/react-router": "^6.0.16",
+        "@testing-library/jest-dom": "^5.16.4",
+        "@testing-library/react": "^13.0.1",
+        "@testing-library/user-event": "^14.1.0",
         "@types/d3": "^7.0.0",
-        "@types/jest": "^26.0.24",
-        "@types/node": "^12.20.16",
-        "@types/qrcode": "^1.4.1",
-        "@types/react": "^16.14.11",
-        "@types/react-dom": "^16.9.14",
-        "@types/react-redux": "^7.1.18",
-        "@types/react-router": "^5.1.16",
-        "@types/react-router-dom": "^5.1.8",
-        "@types/uuid": "^8.3.1",
-        "axios": "^0.21.1",
+        "@types/jest": "^27.4.1",
+        "@types/node": "^17.0.23",
+        "@types/qrcode": "^1.4.2",
+        "@types/react": "^18",
+        "@types/react-dom": "^18",
+        "@types/react-redux": "^7.1.23",
+        "@types/react-router": "^5.1.18",
+        "@types/react-router-dom": "^5.3.3",
+        "@types/uuid": "^8.3.4",
+        "axios": "^0.26.1",
         "customize-cra": "^1.0.0",
         "d3": "^7.0.0",
-        "electron": "^11.4.10",
-        "electron-builder": "^22.11.7",
-        "electron-builder-notarize-pkg": "^1.2.0",
+        "electron": "^16",
+        "electron-builder": "^23.0.4",
+        "electron-builder-notarize-pkg-myh": "^1.3.0",
         "history": "^4.10.1",
-        "ionicons": "^5.5.2",
-        "memfs": "^3.2.0",
-        "qrcode": "^1.4.4",
-        "query-string": "^6.14.1",
-        "react": "^17.0.2",
-        "react-app-rewired": "^2.1.7",
-        "react-dom": "^17.0.2",
-        "react-redux": "^7.2.4",
-        "react-router": "^5.2.0",
-        "react-router-dom": "^5.2.0",
-        "react-scripts": "^4.0.3",
-        "react-test-renderer": "16.2.0",
-        "redux": "^4.1.0",
-        "typescript": "^4.3.5",
-        "uuid": "^8.3.1",
-        "workbox-background-sync": "^5.1.3",
-        "workbox-broadcast-update": "^5.1.3",
-        "workbox-cacheable-response": "^5.1.3",
-        "workbox-core": "^5.1.3",
-        "workbox-expiration": "^5.1.3",
-        "workbox-google-analytics": "^5.1.3",
-        "workbox-navigation-preload": "^5.1.3",
-        "workbox-precaching": "^5.1.3",
-        "workbox-range-requests": "^5.1.3",
-        "workbox-routing": "^5.1.3",
-        "workbox-strategies": "^5.1.3",
-        "workbox-streams": "^5.1.3"
+        "ionicons": "^6",
+        "node-polyfill-webpack-plugin": "^1.1.4",
+        "qrcode": "^1.5.0",
+        "query-string": "^7.1.1",
+        "react": "^18.0.0",
+        "react-app-rewired": "^2.2.1",
+        "react-dom": "^18.0.0",
+        "react-redux": "^7.2.8",
+        "react-router": "^5.2.1",
+        "react-router-dom": "^5.3.0",
+        "react-scripts": "5.0.1",
+        "redux": "^4.1.2",
+        "typescript": "^4.6.3",
+        "uuid": "^8.3.2",
+        "workbox-background-sync": "^5.1.4",
+        "workbox-broadcast-update": "^5.1.4",
+        "workbox-cacheable-response": "^5.1.4",
+        "workbox-core": "^5.1.4",
+        "workbox-expiration": "^5.1.4",
+        "workbox-google-analytics": "^5.1.4",
+        "workbox-navigation-preload": "^5.1.4",
+        "workbox-precaching": "^5.1.4",
+        "workbox-range-requests": "^5.1.4",
+        "workbox-routing": "^5.1.4",
+        "workbox-strategies": "^5.1.4",
+        "workbox-streams": "^5.1.4"
     },
     "eslintConfig": {
         "extends": "react-app"
