@@ -1,5 +1,4 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { BrowserRouter } from "react-router-dom";
 import App, {serviceWorkCallbacks} from './App';
 import Globals from './Globals';
@@ -8,7 +7,7 @@ import * as serviceWorker from './serviceWorkerRegistration';
 // Enable at the very start for logging most messages.
 Globals.enableAppLog();
 
-ReactDOM.render(<BrowserRouter><App /></BrowserRouter>, document.getElementById('root'));
+createRoot(document.getElementById('root')!).render(<BrowserRouter><App /></BrowserRouter>);
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
